Handle redirection correctly for GET reqs
When a redirect is attempted through Turbolinks, it almost works. By
default, a redirect with Go's http.Redirect will serve the correct page
to the frontend via the redirection, but will fail to update the URL.

To fix this, you can send a Turbolinks-Location header in the response
which will make the frontend call its own function to update the URL.
To do this, we needed a way to know that a redirect occured and that we
should in fact send that header, so we store a session value in a cookie
in order to tell the next request it needs to sent the header.