Handle form submission
This adds functionality to handle form submission the way that
Turbolinks and Rails suggest, which is to redirect after form
submission, intercept the redirection, and respond with JavaScript. The
JavaScript calls Turbolinks's visit and clear cache methods in order to
update the frontend correctly, and to correctly update the history API.

The server expects some JavaScript, which is currently in a folder that
is gitignored, but that will be added in the future. If somehow somebody
sees this commit, it's a translation of the coffeescript in Rails itself
under the actionview directory's JavaScript file.

See
https://github.com/rails/rails/blob/master/actionview/app/assets/javascripts/rails-ujs/utils/ajax.coffee
if needed.